fixes to support latest esp32 arduino release.
adc - set gain etc after inlcude ads11115
ESP32HTTPUpdate - old httpupdate no longer works, fixes applied.
maths - float_max - ESP32HTTPUpdate (among others) undef Arduino min and max. easiest fix is to provide my own function.
webpages - toDouble not a string member anymore ? or never was. fixed.
HTTPClient - argument syntax has changed for esp32 lib - fixed.

+++ Arduino/esp32_solar_controller/webpages.ino
@@ -576,7 +576,7 @@ void web_config_submit()
       for(uint8_t x = 0; x < count_cells; x++)
       {
         if (server.argName(i) == String("battery_volt_mod") + String(x+1))
-          config.battery_volt_mod[x] = server.arg(i).toDouble();
+          config.battery_volt_mod[x] = server.arg(i).toFloat();
       }
